server/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
websockets
pydantic
pydantic-settings
